requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
future==0.18.2
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
more-itertools==8.12.0
praw==7.5.0
prawcore==2.3.0
python-dotenv==0.19.2
requests==2.26.0
soupsieve==2.3.1
treelib==1.6.1
update-checker==0.18.0
urllib3==1.26.7
websocket-client==1.2.3
zipp==3.6.0
